+++ docs/fr_FR/Changelog.md
@@ -5,6 +5,7 @@
 -   Amélioration: Suppression cmds obsoletes 'presence.json' & 'presenceEP2.json'.
 -   Amélioration: Support Tuya PIR+24G_radar+light ZG-204ZM (2741).
 -   Amélioration: Outil 'check_json.php' pour subType 'slider'.
+-   Amélioration: Script 'resources/scripts/checkZigate.sh'
 
 ## 250115-BETA-2, ## 250115-BETA-1